+++ src/node-lib/curriculum-api-2023/queries/programmeListing/programmeListing.query.test.ts
@@ -1,6 +1,35 @@
+import {
+  lessonDataFixture,
+  programmeFieldsFixture,
+} from "@oaknational/oak-curriculum-schema";
+
 import sdk from "../../sdk";
 
-import programmeListing from "./programmeListing.query";
+import programmeListing, {
+  getTransformedProgrammeData,
+} from "./programmeListing.query";
+import { programmeListingSchema } from "./programmeListing.schema";
+
+const programmeListingResponse = [
+  {
+    lesson_data: lessonDataFixture(),
+    programme_fields: {
+      ...programmeFieldsFixture(),
+      examboard_display_order: 2,
+    },
+    is_legacy: false,
+    programme_slug: "programme-slug",
+  },
+  {
+    lesson_data: lessonDataFixture(),
+    programme_fields: {
+      ...programmeFieldsFixture(),
+      examboard_display_order: 1,
+    },
+    is_legacy: false,
+    programme_slug: "programme-slug",
+  },
+];
 
 describe("programmeListing()", () => {
   test("throws a not found error if programme is not found", async () => {
@@ -26,34 +55,11 @@ describe("programmeListing()", () => {
           Promise.resolve({
             programmes: [
               {
-                keyStageSlug: "ks3",
-                // subjectSlug: "maths", // <-- missing
-                programmes: [
-                  {
-                    programmeSlug: "maths-secondary-ks4-core",
-                    subjectTitle: "Maths",
-                    unitCount: 61,
-                    lessonCount: 284,
-                    tierSlug: "core",
-                    tierTitle: "Core",
-                  },
-                  {
-                    programmeSlug: "maths-secondary-ks4-foundation",
-                    subjectTitle: "Maths",
-                    unitCount: 63,
-                    lessonCount: 256,
-                    tierSlug: "foundation",
-                    tierTitle: "Foundation",
-                  },
-                  {
-                    programmeSlug: "maths-secondary-ks4-higher",
-                    subjectTitle: "Maths",
-                    unitCount: 63,
-                    lessonCount: 275,
-                    tierSlug: "higher",
-                    tierTitle: "Higher",
-                  },
-                ],
+                lesson_data: {
+                  key_stage: "key-stage",
+                  lesson_id: 4,
+                  lesson_uid: "lesson-uid",
+                },
               },
             ],
           }),
@@ -63,6 +69,105 @@ describe("programmeListing()", () => {
         subjectSlug: "subject-slug",
         isLegacy: false,
       });
-    }).rejects.toThrow(`subjectSlug`);
+    }).rejects.toThrow(`slug`);
+  });
+  test("it returns data in the correct shape", async () => {
+    const res = await programmeListing({
+      ...sdk,
+      programmeListing: jest.fn(() =>
+        Promise.resolve({
+          programmes: programmeListingResponse,
+        }),
+      ),
+    })({
+      keyStageSlug: "ks1",
+      subjectSlug: "maths",
+      isLegacy: false,
+    });
+
+    expect(programmeListingSchema.parse(res)).toEqual({
+      keyStageSlug: "ks1",
+      keyStageTitle: "Key stage 1",
+      programmes: [
+        {
+          examBoardDisplayOrder: 1,
+          examBoardSlug: null,
+          examBoardTitle: null,
+          programmeSlug: "programme-slug",
+          subjectTitle: "subject-description",
+          tierDisplayOrder: null,
+          tierSlug: null,
+          tierTitle: null,
+        },
+        {
+          examBoardDisplayOrder: 2,
+          examBoardSlug: null,
+          examBoardTitle: null,
+          programmeSlug: "programme-slug",
+          subjectTitle: "subject-description",
+          tierDisplayOrder: null,
+          tierSlug: null,
+          tierTitle: null,
+        },
+      ],
+      subjectSlug: "maths",
+      subjectTitle: "Maths",
+    });
+  });
+  test("it sorts by exam board order", async () => {
+    const res = await programmeListing({
+      ...sdk,
+      programmeListing: jest.fn(() =>
+        Promise.resolve({
+          programmes: programmeListingResponse,
+        }),
+      ),
+    })({
+      keyStageSlug: "ks1",
+      subjectSlug: "maths",
+      isLegacy: false,
+    });
+
+    expect(res.programmes[0]?.examBoardDisplayOrder).toBe(1);
+    expect(res.programmes[1]?.examBoardDisplayOrder).toBe(2);
+  });
+  test("getTransformedProgrammeData returns the correct transformed programme data", async () => {
+    const firstProgramme = programmeListingResponse[0];
+    if (!firstProgramme) throw new Error("No first programme");
+    const transformedProgrammes = getTransformedProgrammeData(
+      programmeListingResponse,
+      firstProgramme,
+    );
+
+    expect(transformedProgrammes).toEqual({
+      keyStageSlug: "ks1",
+      keyStageTitle: "Key stage 1",
+      legacy: false,
+      programmes: [
+        {
+          examBoardDisplayOrder: 1,
+          examBoardSlug: null,
+          examBoardTitle: null,
+          programmeSlug: "programme-slug",
+          subjectTitle: "subject-description",
+          tierDisplayOrder: null,
+          tierSlug: null,
+          tierTitle: null,
+        },
+
+        {
+          examBoardDisplayOrder: 2,
+          examBoardSlug: null,
+          examBoardTitle: null,
+          programmeSlug: "programme-slug",
+          subjectTitle: "subject-description",
+          tierDisplayOrder: null,
+          tierSlug: null,
+          tierTitle: null,
+        },
+      ],
+      subjectSlug: "maths",
+      subjectTitle: "Maths",
+    });
   });
 });